[JSONFormConverter] 생성 및 OutputView에 있던 JSON형태로 변환하는 메소드들을 여기로 이동
